TimeMachineView add Sequence BT KinectV2Module working basics

+++ Source/AudioLayerClip.cpp
@@ -66,10 +66,10 @@ void AudioLayerClip::updateAudioSourceFile()
 		clipDuration = reader->lengthInSamples / sampleRate;
 
 		clipLength->setValue(clipDuration);
-		buffer.setSize(reader->numChannels, reader->lengthInSamples);      
-		reader->read(&buffer,                                           
+		buffer.setSize((int)reader->numChannels,(int)reader->lengthInSamples);      
+		reader->read(&buffer,                                       
 			0,                                                      
-			reader->lengthInSamples,                                 
+			reader->lengthInSamples,                                
 			0,                                                      
 			true,                                                    
 			true);
